move the confirm section into its own directory, set the base style for the form and import styles for the buttons
