+++ src/components/Form/Confirmation/Confirmation.styled.tsx
@@ -1,0 +1,4 @@
+import { FormSection } from '@/src/styles/utils/Form.styled'
+import styled from 'styled-components'
+
+export const ConfirmationStyled = styled(FormSection)``
