+++ src/components/Form/Form.tsx
@@ -2,7 +2,7 @@ import { useAppSelector } from '@/src/redux/hooks.tsx'
 import PersonalInfo from './PersonalSection/PersonalSection.tsx'
 import PlanSection from './PlanSection/PlanSection.tsx'
 import AddOnSection from './AddOnSection/AddOnSection.tsx'
-import Confirmation from './Confirmation.tsx'
+import Confirmation from './Confirmation/Confirmation.tsx'
 
 const Form = () => {
   const { step } = useAppSelector((state) => state.form)
